ci: run less e2e jobs when when running from PR (#1239)
## 📜 Description

Reduced amount of iOS e2e tests for PR.

## 💡 Motivation and Context

When we open PR we want to move as quick as possible. Running 2 build
jobs + 6 e2e jobs is a very expensive in terms of time (because we have
only 5 runners available to use simultaneously). So in this PR I decide
to use only 3 jobs for tests:
- iOS 15 (prior to this version only TextKit v1 was used, events were
more synchronous and we had less strange bugs and less strange events
were emitted);
- iOS 17 - classically broken iOS version. On this version we have
TextKit v2, we have early end events etc. +iOS 17 runner is running
fast, so it's our gold standard.
- iOS 26 + Xcode 26 - this is a new combination and we need to assure
that newest iOS runs without issues/regressions.

When we merge into main we run all 6 jobs (just for double
verification).

To achieve this I added a separate step that prepares a matrix for e2e
tests. I've tried other options but this was the only one that was
working 🤷‍♂️

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### CI

- run only 3 iOS e2e for PRs;

## 🤔 How Has This Been Tested?

Tested manually via this PR.

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed